requirements.txt
pycryptodome>=3.23.0
requests>=2.33.0
pyyaml>=6.0.3
APScheduler>=3.11.0
customtkinter>=5.2.2
darkdetect>=0.8.0
